feat: Add Leaflet Draw and Routing Machine types, enhance map functionality and UI components
