+++ roamance/package.json
@@ -51,6 +51,7 @@
     "@eslint/eslintrc": "^3",
     "@shadcn/ui": "^0.0.4",
     "@tailwindcss/postcss": "^4",
+    "@types/leaflet-draw": "^1.0.11",
     "@types/leaflet-routing-machine": "^3.2.8",
     "@types/node": "^20",
     "@types/react": "^19",
