+++ roamance/src/components/maps/LeafletMap.tsx
@@ -6,7 +6,7 @@ import L from 'leaflet';
 import 'leaflet-draw/dist/leaflet.draw.css';
 import 'leaflet-routing-machine/dist/leaflet-routing-machine.css';
 import 'leaflet/dist/leaflet.css';
-import { useRef } from 'react';
+import { useEffect, useRef } from 'react';
 import { MapContainer, TileLayer } from 'react-leaflet';
 
 // Import refactored components
@@ -24,54 +24,58 @@ import { useMapFeatures } from '../../hooks/useMapFeatures';
 import { useMapSearch } from '../../hooks/useMapSearch';
 import { RouteData } from '@/types';
 
-// Initialize Leaflet icons
-delete (L.Icon.Default.prototype as { _getIconUrl?: () => string })._getIconUrl;
-L.Icon.Default.mergeOptions({
-  iconRetinaUrl: '/marker-icon-2x.png',
-  iconUrl: '/marker-icon.png',
-  shadowUrl: '/marker-shadow.png',
-});
+// Initialize Leaflet icons - moved to useEffect to avoid SSR issues
+const initLeafletIcons = () => {
+  delete (L.Icon.Default.prototype as { _getIconUrl?: () => string })._getIconUrl;
+  L.Icon.Default.mergeOptions({
+    iconRetinaUrl: '/marker-icon-2x.png',
+    iconUrl: '/marker-icon.png',
+    shadowUrl: '/marker-shadow.png',
+  });
+};
 
-// Add dark mode styles globally if document is available
-if (typeof document !== 'undefined') {
-  const styleElement = document.createElement('style');
-  styleElement.textContent = `
-    .dark-map .leaflet-tile-pane {
-      filter: brightness(0.95) contrast(1.05) saturate(1.1);
-    }
-    .dark-map .leaflet-control,
-    .dark-map .leaflet-control a {
-      background-color: #1e1e1e !important;
-      color: #f3f4f6 !important;
-      border-color: #374151 !important;
-    }
-    .dark-map .leaflet-control a:hover {
-      background-color: #374151 !important;
-    }
-    .dark-map .leaflet-popup-content-wrapper,
-    .dark-map .leaflet-popup-tip {
-      background-color: #1e1e1e !important;
-      color: #f3f4f6 !important;
-      border-color: #374151 !important;
-    }
-    .leaflet-draw-tooltip {
-      background: rgba(0, 0, 0, 0.5);
-      border: 1px solid rgba(0, 0, 0, 0.5);
-      color: #fff;
-    }
-    .dark-map .leaflet-draw-toolbar a {
-      background-color: #1e1e1e !important;
-      color: #f3f4f6 !important;
-      border-color: #374151 !important;
-    }
-    .dark-map .leaflet-draw-actions a {
-      background-color: #1e1e1e !important;
-      color: #f3f4f6 !important;
-      border-color: #374151 !important;
-    }
-  `;
-  document.head.appendChild(styleElement);
-}
+// Add dark mode styles globally - moved to useEffect to avoid SSR issues
+const addDarkModeStyles = () => {
+  if (typeof document !== 'undefined') {
+    const styleElement = document.createElement('style');
+    styleElement.textContent = `
+      .dark-map .leaflet-tile-pane {
+        filter: brightness(0.95) contrast(1.05) saturate(1.1);
+      }
+      .dark-map .leaflet-control,
+      .dark-map .leaflet-control a {
+        background-color: #1e1e1e !important;
+        color: #f3f4f6 !important;
+        border-color: #374151 !important;
+      }
+      .dark-map .leaflet-control a:hover {
+        background-color: #374151 !important;
+      }
+      .dark-map .leaflet-popup-content-wrapper,
+      .dark-map .leaflet-popup-tip {
+        background-color: #1e1e1e !important;
+        color: #f3f4f6 !important;
+        border-color: #374151 !important;
+      }
+      .leaflet-draw-tooltip {
+        background: rgba(0, 0, 0, 0.5);
+        border: 1px solid rgba(0, 0, 0, 0.5);
+        color: #fff;
+      }
+      .dark-map .leaflet-draw-toolbar a {
+        background-color: #1e1e1e !important;
+        color: #f3f4f6 !important;
+        border-color: #374151 !important;
+      }
+      .dark-map .leaflet-draw-actions a {
+        background-color: #1e1e1e !important;
+        color: #f3f4f6 !important;
+        border-color: #374151 !important;
+      }
+    `;
+    document.head.appendChild(styleElement);
+  }
+};
 
 // Saint Martin Island coordinates as fallback
 export const fallbackLocation = { lat: 20.6295, lng: 92.3208 };
@@ -105,6 +109,14 @@ export default function LeafletMap({
   isCustomStartPoint = false,
   onSearchResultSelect,
 }: LeafletMapProps) {
+  // Initialize client-side only code
+  useEffect(() => {
+    // Initialize Leaflet icons
+    initLeafletIcons();
+    // Add dark mode styles
+    addDarkModeStyles();
+  }, []);
+
   // Use custom hooks for map features and search
   const { searchResults, isSearching, setSearchResults } =
     useMapSearch(searchQuery);
